Add new system and power block diagrams
The complete system/power block diagrams are the ones which a complete version of this project should use.
The updated system/power block diagrams are the ones which my current version of the project uses. The current project does not include the transmission/reception mode written in the project proposal. It only allows users to input Morse code without actually transmitting it. The current version does not use the Wi-Fi dongle.